+++ e3nn_jax/__init__.py
@@ -52,7 +52,7 @@ from e3nn_jax._src.su2 import su2_clebsch_gordan, su2_generators
 from e3nn_jax._src.so3 import clebsch_gordan, generators
 from e3nn_jax._src.instruction import Instruction
 from e3nn_jax._src.irreps import Irrep, MulIrrep, Irreps
-from e3nn_jax._src.irreps_array import IrrepsArray, concatenate, stack, mean, norm, normal, dot
+from e3nn_jax._src.irreps_array import IrrepsArray, concatenate, stack, mean, norm, normal, dot, cross
 from e3nn_jax._src.irreps_array import sum_ as sum
 from e3nn_jax._src.spherical_harmonics import spherical_harmonics, sh, legendre
 from e3nn_jax._src.radial import sus, soft_one_hot_linspace, bessel, poly_envelope, soft_envelope
@@ -142,6 +142,7 @@ __all__ = [
     "norm",
     "normal",
     "dot",
+    "cross",
     "sum",
     "spherical_harmonics",
     "sh",